pgwire,ldapauth: remove ConfigureLDAPAuth hook indirection
Now that ldapauth lives in pkg/security/ldapauth (not under pkg/ccl),
pgwire and authserver can import it directly. This removes the
ConfigureLDAPAuth hook variable, the LDAPManager interface definition
from pgwire, and the noLDAPConfigured stub type.

The LDAPManager interface now lives in the ldapauth package alongside
its implementation. ConfigureLDAPAuth is changed from a package-level
var to a regular function. The "ldap" auth method is registered in
pgwire's loadDefaultMethods() instead of via an init() hook.

The two DB Console tests (TestLDAPAuthorizationDBConsole and
TestLDAPProvisioningDBConsole) are moved to an external test package
(ldapauth_test) to break the import cycle through authserver.
