+++ pkg/ccl/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "//pkg/crosscluster/logical",
         "//pkg/crosscluster/physical",
         "//pkg/crosscluster/producer",
-        "//pkg/security/ldapauth",
         "//pkg/security/oidcauth",
     ],
 )
