+++ pkg/server/authserver/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/jwtauth",
+        "//pkg/security/ldapauth",
         "//pkg/security/password",
         "//pkg/security/provisioning",
         "//pkg/security/username",
