+++ pkg/sql/pgwire/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//pkg/security",
         "//pkg/security/distinguishedname",
         "//pkg/security/jwtauth",
+        "//pkg/security/ldapauth",
         "//pkg/security/password",
         "//pkg/security/provisioning",
         "//pkg/security/sessionrevival",
