+++ pkg/sql/pgwire/auth_methods.go
@@ -16,13 +16,13 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/security/distinguishedname"
 	"github.com/cockroachdb/cockroach/pkg/security/jwtauth"
+	"github.com/cockroachdb/cockroach/pkg/security/ldapauth"
 	"github.com/cockroachdb/cockroach/pkg/security/password"
 	"github.com/cockroachdb/cockroach/pkg/security/provisioning"
 	"github.com/cockroachdb/cockroach/pkg/security/sessionrevival"
 	"github.com/cockroachdb/cockroach/pkg/security/username"
 	"github.com/cockroachdb/cockroach/pkg/server/telemetry"
 	"github.com/cockroachdb/cockroach/pkg/settings"
-	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql"
 	"github.com/cockroachdb/cockroach/pkg/sql/lexbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/hba"
@@ -34,7 +34,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/log/eventpb"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
-	"github.com/cockroachdb/cockroach/pkg/util/uuid"
 	"github.com/cockroachdb/errors"
 	"github.com/cockroachdb/redact"
 	"github.com/go-ldap/ldap/v3"
@@ -86,6 +85,9 @@ func loadDefaultMethods() {
 	// The "trust" method accepts any connection attempt that matches
 	// the current rule.
 	RegisterAuthMethod("trust", authTrust, hba.ConnAny, NoOptionsAllowed)
+
+	// The "ldap" method uses LDAP for authentication.
+	RegisterAuthMethod("ldap", AuthLDAP, hba.ConnAny, ldapauth.CheckHBAEntryLDAP)
 }
 
 // AuthMethod is a top-level factory for composing the various
@@ -916,89 +918,12 @@ func authJwtToken(
 	return b, nil
 }
 
-// LDAPManager is an interface for `ldapauthccl` pkg to add ldap login(authN)
-// and groups sync(authZ) support.
-type LDAPManager interface {
-	// FetchLDAPUserDN extracts the user distinguished name for the sql session
-	// user performing a lookup for the user on ldap server using options provided
-	// in the hba conf and supplied sql username in db connection string.
-	FetchLDAPUserDN(_ context.Context, _ *cluster.Settings,
-		_ username.SQLUsername,
-		_ *hba.Entry,
-		_ *identmap.Conf,
-	) (userDN *ldap.DN, detailedErrorMsg redact.RedactableString, authError error)
-	// ValidateLDAPLogin validates whether the password supplied could be used to
-	// bind to ldap server with the ldap user DN(provided as systemIdentityDN
-	// being the "externally-defined" system identity).
-	ValidateLDAPLogin(_ context.Context, _ *cluster.Settings,
-		_ *ldap.DN,
-		_ username.SQLUsername,
-		_ string,
-		_ *hba.Entry,
-		_ *identmap.Conf,
-	) (detailedErrorMsg redact.RedactableString, authError error)
-	// FetchLDAPGroups retrieves ldap groups for the supplied ldap user
-	// DN(provided as systemIdentityDN being the "externally-defined" system
-	// identity) performing a group search with the options provided in the hba
-	// conf and filtering for the groups which have the user DN as its member.
-	FetchLDAPGroups(_ context.Context, _ *cluster.Settings,
-		_ *ldap.DN,
-		_ username.SQLUsername,
-		_ *hba.Entry,
-		_ *identmap.Conf,
-	) (ldapGroups []*ldap.DN, detailedErrorMsg redact.RedactableString, authError error)
-}
-
 // ldapManager is a singleton global pgwire object which gets initialized from
-// authLDAP method whenever an LDAP auth attempt happens. It depends on ldapauth
-// module to be imported properly to override its default ConfigureLDAPAuth
-// constructor.
+// AuthLDAP method whenever an LDAP auth attempt happens.
 var ldapManager = struct {
 	sync.Once
-	m LDAPManager
+	m ldapauth.LDAPManager
 }{}
-
-type noLDAPConfigured struct{}
-
-func (c *noLDAPConfigured) FetchLDAPUserDN(
-	_ context.Context, _ *cluster.Settings, _ username.SQLUsername, _ *hba.Entry, _ *identmap.Conf,
-) (retrievedUserDN *ldap.DN, detailedErrorMsg redact.RedactableString, authError error) {
-	return nil, "", errors.New("LDAP based authentication requires CCL features")
-}
-
-func (c *noLDAPConfigured) ValidateLDAPLogin(
-	_ context.Context,
-	_ *cluster.Settings,
-	_ *ldap.DN,
-	_ username.SQLUsername,
-	_ string,
-	_ *hba.Entry,
-	_ *identmap.Conf,
-) (detailedErrorMsg redact.RedactableString, authError error) {
-	return "", errors.New("LDAP based authentication requires CCL features")
-}
-
-func (c *noLDAPConfigured) FetchLDAPGroups(
-	_ context.Context,
-	_ *cluster.Settings,
-	_ *ldap.DN,
-	_ username.SQLUsername,
-	_ *hba.Entry,
-	_ *identmap.Conf,
-) (ldapGroups []*ldap.DN, detailedErrorMsg redact.RedactableString, authError error) {
-	return nil, "", errors.New("LDAP based authorization requires CCL features")
-}
-
-// ConfigureLDAPAuth is a hook for the `ldapauth` library to add LDAP login
-// support. It's called to setup the LDAPManager just as it is needed.
-var ConfigureLDAPAuth = func(
-	serverCtx context.Context,
-	ambientCtx log.AmbientContext,
-	st *cluster.Settings,
-	clusterUUID uuid.UUID,
-) LDAPManager {
-	return &noLDAPConfigured{}
-}
 
 // AuthLDAP is the AuthMethod constructor for the CRDB-specific ldap auth
 // mechanism. The "LDAP" method requires a clear text password which will be
@@ -1018,7 +943,7 @@ func AuthLDAP(
 ) (*AuthBehaviors, error) {
 	ldapManager.Do(func() {
 		if ldapManager.m == nil {
-			ldapManager.m = ConfigureLDAPAuth(sCtx, execCfg.AmbientCtx, execCfg.Settings, execCfg.NodeInfo.LogicalClusterID())
+			ldapManager.m = ldapauth.ConfigureLDAPAuth(sCtx, execCfg.AmbientCtx, execCfg.Settings, execCfg.NodeInfo.LogicalClusterID())
 		}
 	})
 	b := &AuthBehaviors{}
@@ -1149,7 +1074,7 @@ func AuthLDAP(
 func AuthorizeLDAPHelper(
 	ctx context.Context,
 	b *AuthBehaviors,
-	ldapManager LDAPManager,
+	ldapManager ldapauth.LDAPManager,
 	execCfg *sql.ExecutorConfig,
 	ldapUserDN *ldap.DN,
 	user username.SQLUsername,
